More new authentication code.
