+++ awsraw.gemspec
@@ -21,7 +21,7 @@ Gem::Specification.new do |s|
 
   # specify any dependencies here; for example:
   s.add_development_dependency "rake"
-  s.add_development_dependency "rspec"
+  s.add_development_dependency "rspec", "~> 2.14"
   s.add_runtime_dependency "finer_struct", "~> 0.0.5"
   s.add_runtime_dependency "faraday", "~> 0.8.8"
 end
